% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma_utils.R
\name{calculate_control_limit_ewma}
\alias{calculate_control_limit_ewma}
\title{Calibrating Control Limits for EWMA Control Charts}
\usage{
calculate_control_limit_ewma(
  lambda,
  ARL = 500L,
  fir = TRUE,
  side = c("both", "upper", "lower"),
  replications = 5000L,
  interval = c(1e-04, 3.5),
  distribution_fun = stats::rnorm,
  ...
)
}
\arguments{
\item{lambda}{\code{numeric}: The smoothing parameter, lower values indicate stronger smoothing and is associated with earlier detection of smaller shifts.
Larger shifts are typically detected earlier with higher values of lambda. Must lie in the interval \verb{(0,1]}. \code{lambda=1} results in a shewart control chart.}

\item{ARL}{\code{integer}: The desired ARL, defaults to 500.}

\item{fir}{\code{logical}: should Fast Initial Response (FIR) be used? If \code{TRUE}, the initial value of the EWMA is set to the first value of the data.}

\item{side}{\code{character}: The side of the control chart, one of "both", "upper", "lower".}

\item{replications}{\code{integer}:Number of replications used in the, Monte-Carlo simulation.}

\item{interval}{\verb{vector<numeric>}: The interval over which the control limit is searched.}

\item{distribution_fun}{\code{function}: Distribution function that is used in the simulation. Defaults to \code{\link[stats:Normal]{stats::rnorm()}}, but user-specified
functions can be provided (see details).}

\item{...}{\code{ANY}: Additional arguments that will be passed on to \code{distribution_fun}.}
}
\value{
\code{numeric}: The control limit h.
}
\description{
Calculates the critical value \code{h} governing the control limits for EWMA Control Charts. \code{h} is chosen such that the observed
Average Run Length (ARL) of the process is as close as possible to the user-specified ARL. The value is determined through numerical optimization
over a Monte-Carlo simulation.
}
\details{
Usually, the time-series is assumed to be iid normally distributed. However, different distribution functions can be used.
To this end, an arbitrary function can be provided for \code{distribution_fun} and the control limits will be chosen accordingly. Additional parameters can be supplied by \code{...}.
\strong{Note:} The function must fulfill two properties:
\itemize{
\item It must have a parameter \code{n} that governs the sample size.
\item The resulting process has mean zero
}
}
\examples{
# normally distributed observations
calculate_control_limit_ewma(
  lambda = 0.05, ARL = 250L, fir = TRUE, side = "both",
  replications = 500L, distribution_fun = stats::rnorm
)
# t-distributed observations with additional parameter
calculate_control_limit_ewma(
  lambda = 0.05, ARL = 250L, fir = TRUE, side = "both",
  replications = 500L, distribution_fun = stats::rt, df = 3
)

}
