% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma_utils.R
\name{cl}
\alias{cl}
\title{Construct Control Limits for EWMA Control Charts (internal)}
\usage{
cl(lambda, h, T, fir = TRUE)
}
\arguments{
\item{lambda}{\code{numeric}: The smoothing parameter, lower values indicate stronger smoothing and is associated with earlier detection of smaller shifts.
Larger shifts are typically detected earlier with higher values of lambda. Must lie in the interval \verb{(0,1]}. \code{lambda=1} results in a shewart control chart.}

\item{h}{\code{numeric}: A positive numeric used for constructing the control limit. Must be typically chosen together with \code{lambda}.}

\item{T}{\code{integer}: Desired length of control limit vector}

\item{fir}{\code{logical}: should Fast Initial Response (FIR) be used? If \code{TRUE}, the initial value of the EWMA is set to the first value of the data.}
}
\value{
\verb{vector<numeric>}: control limits
}
\description{
Constructs control limits for EWMA control charts and a given critical value \code{h}. They are always one-sided and positive,
so they must be transformed, if they should be applied to lower or two-sided control charts.
}
