% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlcharts-package.R
\docType{package}
\name{controlcharts-package}
\alias{controlcharts}
\alias{controlcharts-package}
\title{controlcharts: Interactive Plotting for Funnel and Statistical Process Control Charts}
\description{
An 'R' front-end for the WA Health 'PowerBI' custom visuals for funnel plots and SPC charts.
}
\author{
\strong{Maintainer}: Andrew R. Johnson \email{andrew.johnson@arjohnsonau.com} (\href{https://orcid.org/0000-0001-7000-8065}{ORCID})

}
\keyword{internal}
