% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma.R
\name{ewma_cc}
\alias{ewma_cc}
\title{Exponentially Weighted Moving Average (EWMA) Control Chart}
\usage{
ewma_cc(
  x,
  lambda = 0.05,
  h,
  side = c("both", "upper", "lower"),
  fir = TRUE,
  stop_on_exceedance = FALSE
)
}
\arguments{
\item{x}{\verb{vector<numeric>}: A numeric vector of data with mean zero and unit variance.}

\item{lambda}{\code{numeric}: The smoothing parameter, lower values indicate stronger smoothing and is associated with earlier detection of smaller shifts.
Larger shifts are typically detected earlier with higher values of lambda. Must lie in the interval \verb{(0,1]}. \code{lambda=1} results in a shewart control chart.}

\item{h}{\code{numeric}: A positive numeric used for constructing the control limit. Must be typically chosen together with \code{lambda}.}

\item{side}{\code{character}: The side of the control chart, one of "both", "upper", "lower".}

\item{fir}{\code{logical}: should Fast Initial Response (FIR) be used? If \code{TRUE}, the initial value of the EWMA is set to the first value of the data.}

\item{stop_on_exceedance}{\code{logical}: If \code{TRUE}, the function stops when the control limit is exceeded. Otherwise processes x completely.}
}
\value{
\code{list}: A list with the following components:
\itemize{
\item \verb{z vector<numeric>} - The EWMA statistic.
\item \verb{upper_limit/lower_limit vector<numeric>} - Vector, containing the control limits.
\item \verb{exceedance vector<logical>} - TRUE at times where exceedances are detected FALSE otherwise.
\item \code{side }character`` - The side of the control chart.
}
}
\description{
This function calculates the exponentially weighted moving average (EWMA) control chart.
}
\examples{
res <- ewma_cc(x = rnorm(2e3), lambda = 0.05, h = 2.615)


}
