% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ewma_smoothing}
\alias{ewma_smoothing}
\title{Smoothing using exponential weighted moving average}
\usage{
ewma_smoothing(x, lambda)
}
\arguments{
\item{x}{\verb{vector<numeric>}: A numeric vector of data with mean zero and unit variance.}

\item{lambda}{\code{numeric}: The smoothing parameter, lower values indicate stronger smoothing and is associated with earlier detection of smaller shifts.
Larger shifts are typically detected earlier with higher values of lambda. Must lie in the interval \verb{(0,1]}. \code{lambda=1} results in a shewart control chart.}
}
\value{
\code{vector<numeric}: The smoothed series
}
\description{
Smoothing using exponential weighted moving average
}
