% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{funnel_default_settings}
\alias{funnel_default_settings}
\title{Get default settings for Funnel charts
Retrieve the default settings for Funnel charts or a specific settings group.}
\usage{
funnel_default_settings(group = NULL)
}
\arguments{
\item{group}{Optional. A specific settings group to retrieve. If NULL, all settings groups are returned.}
}
\value{
A list of default settings for Funnel charts or the specified settings group.
}
\description{
Get default settings for Funnel charts
Retrieve the default settings for Funnel charts or a specific settings group.
}
\examples{
#' # Get all default settings for Funnel charts
funnel_default_settings()
# # Get default settings for a specific group
funnel_default_settings("x_axis")
}
