% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc-shiny}
\alias{misc-shiny}
\alias{miscOutput}
\alias{renderMisc}
\title{Shiny bindings for multi-indicator sigma charts}
\usage{
miscOutput(outputId, width = "100\%", height = "400px")

renderMisc(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Valid CSS dimensions.}

\item{expr}{An expression that generates a MISC chart.}

\item{env}{Environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is quoted.}
}
\value{
An interactive Shiny widget.
}
\description{
Shiny bindings for multi-indicator sigma charts
}
