% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{misc_default_settings}
\alias{misc_default_settings}
\title{Get default settings for multi-indicator sigma charts}
\usage{
misc_default_settings(group = NULL)
}
\arguments{
\item{group}{Optional settings group. If \code{NULL}, all groups are returned.}
}
\value{
A list of MISC settings.
}
\description{
Get default settings for multi-indicator sigma charts
}
