% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma_utils.R
\name{optim_func_ewma}
\alias{optim_func_ewma}
\title{Sub-Function for optimization over the control limit critical values (internal)}
\usage{
optim_func_ewma(h, ARL, z_mat, T, exceedance, cl_prelim, fallback_arl)
}
\arguments{
\item{h}{\code{numeric}: Critical value for the control limit}

\item{ARL}{\code{integer}: Desired ARL}

\item{z_mat}{\verb{matrix<numeric>}: EWMA smoothed data. Dimensions: \code{T} x replications}

\item{T}{\code{integer}: Sample length}

\item{exceedance}{\code{function}: Function, that checks for breaches of the control limit}

\item{cl_prelim}{\code{numeric}: preliminary control limit, i.e. needs to be scaled by \code{h}}

\item{fallback_arl}{\code{integer}: If no breach is detected, which (artificial) ARL should be returned?}
}
\value{
\code{numeric}: Squared difference between desired and actual ARL
}
\description{
Is called by \code{\link[=calculate_control_limit_ewma]{calculate_control_limit_ewma()}}.
}
\seealso{
calculate_control_limit_ewma
}
