% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chart.R
\name{plot_ewma}
\alias{plot_ewma}
\title{Erstellt die Grafik fuer einen EWMA Control Chart}
\usage{
plot_ewma(ewma_cc, dates = NULL, plot_name = "EWMA Control Chart")
}
\arguments{
\item{ewma_cc}{\code{list}: The value of the function \link{ewma_cc}}

\item{dates}{\verb{vector<dates>}: Optional. The dates corresponding to each observation. Must be of the same length as the control chart.}

\item{plot_name}{\code{character}: The title to be displayed above the plot. Defaults to 'EWMA Control Chart'}
}
\value{
If ggplot2 is installed, a plot. Otherwise NULL
}
\description{
Creates a plot for an EWMA control chart. If available, \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} will be used, otherwise, the plot will be created with base R.
}
