% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.static_plot}
\alias{print.static_plot}
\title{Method to print a static plot to the R graphics device
Uses the QuickJSR package for running the visual JS to create an SVG string
of the plot, then the rsvg package to render that SVG to the graphics device}
\usage{
\method{print}{static_plot}(x, ...)
}
\description{
Method to print a static plot to the R graphics device
Uses the QuickJSR package for running the visual JS to create an SVG string
of the plot, then the rsvg package to render that SVG to the graphics device
}
