% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{spc-limits}
\alias{spc-limits}
\alias{spc_limits}
\title{Calculate control limits}
\usage{
spc_limits(
  keys,
  numerators,
  denominators,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Calculate control limits
}
