% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{spc_default_settings}
\alias{spc_default_settings}
\title{Get default settings for SPC charts}
\usage{
spc_default_settings(group = NULL)
}
\arguments{
\item{group}{Optional. A specific settings group to retrieve.
If NULL, all settings groups are returned.}
}
\value{
A list of default settings for SPC charts or the specified
settings group.
}
\description{
Retrieve the default settings for SPC charts or a specific settings group.
}
\examples{
#' # Get all default settings for SPC charts
spc_default_settings()
# # Get default settings for a specific group
spc_default_settings("x_axis")
}
