% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.R
\name{funnel}
\alias{funnel}
\title{Generate interactive funnel chart}
\usage{
funnel(
  keys,
  numerators,
  denominators,
  data,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Generate interactive funnel chart
}
