% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.R
\name{funnel}
\alias{funnel}
\title{Generate interactive Funnel chart}
\usage{
funnel(
  data,
  keys,
  numerators,
  denominators,
  tooltips,
  labels,
  aggregations = list(numerators = "sum", denominators = "sum", tooltips = "first",
    labels = "first"),
  canvas_settings = NULL,
  funnel_settings = NULL,
  outlier_settings = NULL,
  scatter_settings = NULL,
  line_settings = NULL,
  x_axis_settings = NULL,
  y_axis_settings = NULL,
  label_settings = NULL,
  tooltip_settings = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A data frame containing the data for the chart.}

\item{keys}{A vector or column name representing the categories of the chart.}

\item{numerators}{A numeric vector or column name representing the numerators for each category.}

\item{denominators}{A numeric vector or column name representing the denominators for each category.}

\item{tooltips}{A vector or column name representing the tooltips for each category.}

\item{labels}{A vector or column name representing the labels for each category.}

\item{aggregations}{A list of aggregation function names for each field if multiple values are provided for each key. Valid options are:
\itemize{
\item \code{"first"}: returns the first value
\item \code{"last"}: returns the last value
\item \code{"sum"}: returns the sum of values
\item \code{"mean"}: returns the mean of values
\item \code{"min"}: returns the minimum value
\item \code{"max"}: returns the maximum value
\item \code{"median"}: returns the median value
\item \code{"count"}: returns the count of values
}}

\item{canvas_settings}{Optional list of settings for the canvas, see \code{funnel_default_settings('canvas')} for valid options.}

\item{funnel_settings}{Optional list of settings for the Funnel chart, see \code{funnel_default_settings('funnel')} for valid options.}

\item{outlier_settings}{Optional list of settings for outliers, see \code{funnel_default_settings('outliers')} for valid options.}

\item{scatter_settings}{Optional list of settings for scatter points, see \code{funnel_default_settings('scatter')} for valid options.}

\item{line_settings}{Optional list of settings for lines, see \code{funnel_default_settings('lines')} for valid options.}

\item{x_axis_settings}{Optional list of settings for the x-axis, see \code{funnel_default_settings('x_axis')} for valid options.}

\item{y_axis_settings}{Optional list of settings for the y-axis, see \code{funnel_default_settings('y_axis')} for valid options.}

\item{label_settings}{Optional list of settings for labels, see \code{funnel_default_settings('labels')} for valid options.}

\item{tooltip_settings}{Optional list of settings for tooltips, see \code{spc_default_settings('tooltips')} for valid options.}

\item{width}{Optional width of the chart in pixels. If NULL (default), the chart will fill the width of its container.}

\item{height}{Optional height of the chart in pixels. If NULL (default), the chart will fill the height of its container.}

\item{elementId}{Optional HTML element ID for the chart.}
}
\value{
An object of class \code{controlchart} containing the interactive plot, static plot, limits data frame, raw data, and a function to save the plot.
}
\description{
Generate interactive Funnel chart
}
