% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{spc-limits}
\alias{spc-limits}
\alias{spc_limits}
\title{Calculate control limits}
\usage{
spc_limits(
  keys = NULL,
  numerators = NULL,
  denominators = NULL,
  rebaseline_groupings = NULL,
  xbar_sds = NULL,
  width = NULL,
  height = NULL
)
}
\description{
Calculate control limits
}
