% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{spc}
\alias{spc}
\title{Generate interactive SPC chart}
\usage{
spc(
  keys,
  numerators,
  denominators,
  data,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Generate interactive SPC chart
}
